Move VhostUserMsgHeader to a separate file
